refactor(js/plugins/vertexai): extract common logic
