+++ js/plugins/vertexai/package.json
@@ -55,7 +55,7 @@
     "@types/node": "^20.11.16",
     "npm-run-all": "^4.1.5",
     "rimraf": "^6.0.1",
-    "tsup": "^8.0.2",
+    "tsup": "6.5.0",
     "tsx": "^4.7.0",
     "typescript": "^4.9.0"
   },
@@ -66,6 +66,11 @@
       "import": "./lib/index.mjs",
       "types": "./lib/index.d.ts",
       "default": "./lib/index.js"
+    }
+  },
+  "typesVersions": {
+    "*": {
+
     }
   }
 }
